refactor aks tests
